models/pipelines/fashion/requirements-prediction.txt
starlette
google-cloud-storage
dill==0.3.4
scikit-learn==0.24.2
uvicorn